Linted new components
